move toggleMap to layout reducer
