+++ client/src/filter/spatial/MapFilterContainer.js
@@ -1,15 +1,13 @@
 import { connect } from 'react-redux'
 import MapFilter from './MapFilter'
 import { toggleExcludeGlobal } from '../../actions/SearchParamActions'
-import { toggleMap, updateBounds } from '../../actions/FlowActions'
+import { toggleMap } from '../../actions/LayoutActions'
 import { clearCollections, triggerSearch } from "../../actions/SearchRequestActions"
 import { showCollections } from "../../actions/FlowActions"
 
 const mapStateToProps = state => {
   return {
-    showMap: state.ui.mapFilter.showMap,
-    bounds: state.ui.mapFilter.bounds,
-    boundsSource: state.ui.mapFilter.boundsSource,
+    showMap: state.ui.layout.showMap,
     geoJSON: state.behavior.search.geoJSON
   }
 }
@@ -27,9 +25,6 @@ const mapDispatchToProps = dispatch => {
     toggleMap: () => {
       dispatch(toggleMap())
     },
-    updateBounds: (bounds, source) => {
-      dispatch(updateBounds(bounds, source))
-    }
   }
 }
