+++ client/src/reducers/reducer.js
@@ -13,7 +13,6 @@ import cardDetails from './ui/cardDetails'
 import granuleDetails from './ui/granuleDetails'
 import loading from './ui/loading'
 import background from './ui/background'
-import mapFilter from './ui/mapFilter'
 import layout from './ui/layout'
 
 const domain = combineReducers({
@@ -27,7 +26,6 @@ const ui = combineReducers({
   granuleDetails,
   loading,
   background,
-  mapFilter,
   layout
 })
